Defined point and vector constructor methods
Points and Vectors are both tuples under the hood with the w parameter defining the difference. The String method explicitly tells you which is which.

First test for the point function does not actually work. I think it may have something to do with the way the fact that the point method returns a pointer to a tuple.... looking into this